Refactored crews code to make it more modular. Added more API endpoints to make the crews page more streamlined.
