+++ server/src/routes/users.js
@@ -24,66 +24,143 @@ router.get("/", (req, res) => {
 // });
 
 
-router.get("/schedule", async (req, res) => {
-    let data = []
+// router.get("/schedule", async (req, res) => {
+//     let data = []
 
-    try{
-        let courseDates = await knex("users")
+//     try{
+//         let courseDates = await knex("users")
+//         .join('course_registration', 'users.id', 'user_id')
+//         .join('courses', 'courses.id', 'course_id')
+//         .select('users.id as user_id','first_name','last_name','course_registration.id as registration_id' ,'course_id','course_name', 'cert_granted', 'date_start', 'date_end')
+//         .select(knex.raw(`'courses' as source`))
+//         .orderBy('users.id');
+
+//         let crewDates = await knex("users")
+//         .join('crews', 'users.crew_id', 'crews.id')
+//         .join('crew_rotations', 'crews.id', 'crew_rotations.crew_id')
+//         .select('users.id as user_id','first_name', 'last_name','crews.id as crew_id', 'crew_name','crew_rotations.id as rotation_id','shift_type',  'date_start', 'date_end')
+//         .select(knex.raw(`'crews' as source`))
+//         .orderBy('users.id');
+
+//         data.push({crewDates})
+//         data.push({courseDates})
+
+//         return res.status(200).json(data)
+//     }
+//     catch (error){
+//         return res.status(500).json({ error: error });
+//     }
+// });
+
+router.get("/schedule", async (req, res) => {
+    try {
+      const courseDates = await knex("users")
         .join('course_registration', 'users.id', 'user_id')
         .join('courses', 'courses.id', 'course_id')
-        .select('users.id as user_id','first_name','last_name','course_registration.id as registration_id' ,'course_id','course_name', 'cert_granted', 'date_start', 'date_end')
-        .select(knex.raw(`'courses' as source`))
-        .orderBy('users.id');
+        .select('users.id as user_id', 'date_start', 'date_end');
 
-        let crewDates = await knex("users")
+      const crewDates = await knex("users")
         .join('crews', 'users.crew_id', 'crews.id')
         .join('crew_rotations', 'crews.id', 'crew_rotations.crew_id')
-        .select('users.id as user_id','first_name', 'last_name','crews.id as crew_id', 'crew_name','crew_rotations.id as rotation_id','shift_type',  'date_start', 'date_end')
-        .select(knex.raw(`'crews' as source`))
-        .orderBy('users.id');
+        .select('users.id as user_id', 'date_start', 'date_end');
 
-        data.push({crewDates})
-        data.push({courseDates})
+      const combined = [
+        ...courseDates.map(d => ({ ...d, type: 'course' })),
+        ...crewDates.map(d => ({ ...d, type: 'crew' }))
+      ];
 
-        return res.status(200).json(data)
+      const grouped = {};
+
+      for (const entry of combined) {
+        const { user_id, date_start, date_end, type } = entry;
+
+        if (!grouped[user_id]) {
+          grouped[user_id] = {
+            user_id,
+            dates: []
+          };
+        }
+
+        grouped[user_id].dates.push({ date_start, date_end, type });
+      }
+
+      const result = Object.values(grouped);
+
+      return res.status(200).json(result);
+    } catch (error) {
+      console.error(error);
+      return res.status(500).json({ error: "Internal server error" });
     }
-    catch (error){
-        return res.status(500).json({ error: error });
-    }
-});
+  });
+
 
 // get route to get the schedule of a specific user
+// router.get("/schedule/:id", async (req, res) => {
+//     const id = parseInt(req.params.id)
+//     let data = []
+
+//     try{
+//         let courseDates = await knex("users")
+//         .join('course_registration', 'users.id', 'user_id')
+//         .join('courses', 'courses.id', 'course_id')
+//         .select('users.id as user_id','first_name','last_name','course_registration.id as registration_id' ,'course_id','course_name', 'cert_granted', 'date_start', 'date_end')
+//         .select(knex.raw(`'courses' as source`))
+//         .orderBy('course_id')
+//         .where('users.id', id);
+
+//         let crewDates = await knex("users")
+//         .join('crews', 'users.crew_id', 'crews.id')
+//         .join('crew_rotations', 'crews.id', 'crew_rotations.crew_id')
+//         .select('users.id as user_id','first_name', 'last_name','crews.id as crew_id', 'crew_name','crew_rotations.id as rotation_id','shift_type',  'date_start', 'date_end', 'shift_duration', 'role')
+//         .select(knex.raw(`'crews' as source`))
+//         .orderBy('rotation_id')
+//         .where('users.id', id);
+
+//         data.push({crewDates})
+//         data.push({courseDates})
+
+//         return res.status(200).json(data)
+//     }
+//     catch (error){
+//         return res.status(500).json({ error: error });
+//     }
+
+// });
+
 router.get("/schedule/:id", async (req, res) => {
     const id = parseInt(req.params.id)
-    let data = []
 
     try{
         let courseDates = await knex("users")
         .join('course_registration', 'users.id', 'user_id')
         .join('courses', 'courses.id', 'course_id')
-        .select('users.id as user_id','first_name','last_name','course_registration.id as registration_id' ,'course_id','course_name', 'cert_granted', 'date_start', 'date_end')
-        .select(knex.raw(`'courses' as source`))
-        .orderBy('course_id')
+        .select('date_start', 'date_end')
         .where('users.id', id);
 
         let crewDates = await knex("users")
         .join('crews', 'users.crew_id', 'crews.id')
         .join('crew_rotations', 'crews.id', 'crew_rotations.crew_id')
-        .select('users.id as user_id','first_name', 'last_name','crews.id as crew_id', 'crew_name','crew_rotations.id as rotation_id','shift_type',  'date_start', 'date_end', 'shift_duration', 'role')
-        .select(knex.raw(`'crews' as source`))
-        .orderBy('rotation_id')
+        .select('date_start', 'date_end')
         .where('users.id', id);
 
-        data.push({crewDates})
-        data.push({courseDates})
+        let combinedDates = {
+            user_id: id,
+            dates: [
+              ...courseDates.map(d => ({ ...d, type: 'course' })),
+              ...crewDates.map(d => ({ ...d, type: 'crew' }))
+            ]
+          };
 
-        return res.status(200).json(data)
+
+
+        return res.status(200).json(combinedDates)
     }
     catch (error){
         return res.status(500).json({ error: error });
     }
 
 });
+
 
 
 router.get("/:id", async (req, res) => {
